+++ packages/core/lib/combo/settings/data.ts
@@ -1662,12 +1662,12 @@ export const SETTINGS = [{
   cond: hasMM,
 }, {
   key: 'bronzeScale',
-  name: 'Bronze Scale (WIP)',
+  name: 'Bronze Scale (EXPERIMENTAL)',
   category: 'items.extensions',
   type: 'boolean',
-  description: 'Add an extra scale item, that is required to enter deep water. WORK IN PROGRESS, MIGHT BREAK.',
+  description: 'Add an extra scale item, that is required to enter deep water. EXPERIMENTAL LOGIC - NO LOGIC FOR MQ.',
   default: false,
-  cond: (s: any) => (hasOoT(s) || s.scalesMm) && s.logic === 'none', /* TODO: Enable once logic is ready */
+  cond: (s: any) => (hasOoT(s) || s.scalesMm),
 }, {
   key: 'childWallets',
   name: 'Child Wallets',